+++ src/kiro/model/credentials.rs
@@ -197,6 +197,7 @@ impl KiroCredentials {
 }
 
 #[cfg(test)]
+#[allow(clippy::field_reassign_with_default)]
 mod tests {
     use super::*;